Bugs corrigidos e rotina de testes feita
