--- INV_PROJV1/Suppliers.csv
@@ -1,3 +1,0 @@
-Produto,Tipo,Modelo,Cor,Tamanho,Quantidade
-Tira,Lisa,Aviador,Branca,22,10
-Borracha,90/10,Comum,Azul/Vermelho,90cm,2
