+++ Products.csv
@@ -1,0 +1,10 @@
+Produto,Tipo,Modelo,Cor,Tamanho,Quantidade
+Tira,Lisa,Aviador,Branca,22,10
+Borracha,90/10,Comum,Azul/Vermelho,90cm,2
+Símbolo,Para colar,Logomarca,Padrão fundo preto,Regular,50
+Forma,Matálica,Padrão,Azul,23/24,1
+Forma,Metálica,Padrão,Azul,25/26,1
+Forma,Metálica,Padrão,Azul,27/28,1
+Forma,Metálica,Padrão,Azul,29/30,1
+Tira,Texturizada,Comum,Preta,29,10
+Borrachaa,80/20,Comum,,90cm,2
