--- teste.csv
@@ -1,4 +1,0 @@
-Produto,Tipo,Modelo,Cor,Tamanho,Quantidade
-Tira,Lisa,Aviador,Branca,22,10
-Forma,Metálica,Regular,--,22,1
-Borracha,Regularr,Comum,Azul/Vermelho,90cm,2
